Add structure and necessary files CSS practice
Created folders pages and styles. Moved the initial index.html to the
pages folder and the initial styles.css to the styles folder. Renamed
that index.html file to 01-css-methods.html to match the name of the
practice. Renamed styles.css to 01-css-methods.css to match the name of
the html file.

Created a new index.html that acts as a home page and links to the
each of the 5 practices.

Created the rest of the html files for the practice within pages folder
with its respective CSS file within the styles folder.