Implement 02-margin-and-padding
With this exercise, didn't look at the self-check and decided to
implement it by myself, and it was a big struggle. I didn't know that to
align the text, I could've used text-align, since it's been hammered in
that I can use margin: auto to horizontally align elements. I also
didn't know the specific padding / margin pixels that I was supposed to
use, and I failed to see that I had to change specific margins such as
margin-top as opposed to a whole margin.

I hope I get more practice with this stuff so that I can be more
confident in my CSS abilities.